requirement.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==4.2.14
django-bootstrap-v5==1.0.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-extensions==3.2.3
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
importlib_metadata==8.2.0
kombu==5.3.7
Markdown==3.6
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.19.2
